Refactoring, cleanup and generalization
