+++ lib/neo4j/core/cypher_session.rb
@@ -1,7 +1,10 @@
+require 'active_support'
+
 module Neo4j
   module Core
     class CypherSession
       attr_reader :adaptor
+      delegate :close, to: :adaptor
 
       def initialize(adaptor)
         fail ArgumentError, "Invalid adaptor: #{adaptor.inspect}" if !adaptor.is_a?(Adaptors::Base)
